Step 5: Move PublicRegisterRenderTest and PublicRegisterReferralTest to core
PublicRegisterView now lives in danceschool.core, so its tests belong there
too. Classes extracted from danceschool/register/tests.py (git history at
that path) and appended to danceschool/core/tests.py with:

- _url() updated: reverse('registration') → reverse('publicRegistration')
- Referral URL names updated: registrationWithVoucher →
  publicRegistrationWithVoucher, registrationWithMarketingId →
  publicRegistrationWithMarketingId
- Import updated: from .models import PublicRegisterEventPluginChoice →
  from danceschool.core.models import PublicRegisterEventPluginChoice
- Added `from cms.api import add_plugin` to core/tests.py imports

Removed the two classes and now-unused imports (json, Invoice, updateConstant,
REG_VALIDATION_STR, PublicEvent) from register/tests.py.
